requirements.txt
gcloud
pandas
pyyaml
